pyproject.toml
[project]
name = "saas-fastapi-example"
version = "0.1.0"
description = "SaaS fastapi example application"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "anyio>=4.0.0",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "black>=26.1.0",
    "fastapi[standard]>=0.128.8",
    "mypy>=1.19.1",
    "passlib>=1.7.4",
    "pre-commit>=4.5.1",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "pytest>=9.0.2",
    "sqlalchemy>=2.0.46",
    "sqlmodel>=0.0.33",
    "types-passlib>=1.7.7.20260211",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["app/tests"]

[tool.isort]
profile = "black"
known_first_party = ["app"]
skip = [".venv", "venv"]
